Change datatype to int16

+++ record_audio.py
@@ -5,7 +5,7 @@ import scipy.io.wavfile as wav
 def record():
     fs=16000 # sampling rate
     duration = 5  # seconds
-    audio = sd.rec(duration * fs, samplerate=fs, channels=1,dtype='float32')
+    audio = sd.rec(duration * fs, samplerate=fs, channels=1,dtype='int16')
     print "Recording Audio"
     sd.wait()
     print "Audio recording complete"
